Add Sorter Functionality And Clean Code
Create the various components for adding the
sorting functionality to the table. A little
code cleaning was done with PHP CS.

Pop Up modal component was also added to present
details to the user on clicking of the table row
All components for this has been created and added
accordingly.